+++ src/components/super/i-data/i-data.ts
@@ -95,7 +95,14 @@ export default abstract class iData extends iDataHandlers {
 
 			const setDBData = (data: CanUndef<this['DB']>) => {
 				this.saveDataToRootStore(data);
-				this.hydrationStore?.set(this.componentId, providerHydrationKey, Object.cast(data));
+
+				if (data !== undefined) {
+					this.hydrationStore?.set(this.componentId, providerHydrationKey, Object.cast(data));
+
+				} else {
+					this.hydrationStore?.setEmpty(this.componentId, providerHydrationKey);
+				}
+
 				this.db = this.convertDataToDB<this['DB']>(data);
 
 				// During hydration, there may be a situation where the cache on the DB getter is set before rendering occurs,
@@ -257,12 +264,10 @@ export default abstract class iData extends iDataHandlers {
 		return super.initLoad(data, opts);
 	}
 
-	override reload(opts?: InitLoadOptions): Promise<void> {
-		if (!this.r.isOnline && !this.offlineReload) {
-			return Promise.resolve();
+	override async reload(opts?: InitLoadOptions): Promise<void> {
+		if ((await this.remoteState.net.isOnline()).status || this.offlineReload) {
+			return super.reload(opts);
 		}
-
-		return super.reload(opts);
 	}
 
 	/**
@@ -280,7 +285,7 @@ export default abstract class iData extends iDataHandlers {
 			return;
 		}
 
-		this.r.providerDataStore.set(key, data);
+		this.r.providerDataStore?.set(key, data);
 
 		function getKey(val: string | CanUndef<iData['dataProviderProp']>): CanUndef<string> {
 			if (val == null || Object.isString(val)) {