+++ src/components/super/i-page/interface.ts
@@ -14,6 +14,9 @@ export type TitleValue<CTX extends iPage = iPage['unsafe']> =
 
 export type DescriptionValue<CTX extends iPage = iPage['unsafe']> = TitleValue<CTX>;
 
+export type TitleValueProp<CTX extends iPage = iPage['unsafe']> = TitleValue<CTX> | (() => TitleValue<CTX>);
+export type DescriptionValueProp <CTX extends iPage = iPage['unsafe']> = TitleValueProp<CTX>;
+
 export interface StageTitles<CTX extends iPage = iPage['unsafe']> extends Dictionary<TitleValue<CTX>> {
 	'[[DEFAULT]]': TitleValue<CTX>;
 }