+++ src/core/component/hydration/const.ts
@@ -6,4 +6,10 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-export const styles = new Map<string, Promise<{default: string}>>();
+export const
+	styles = new Map<string, Promise<{default: string}>>();
+
+/**
+ * A key for empty data in the hydration storage
+ */
+export const emptyDataStoreKey = '[[EMPTY]]';