+++ src/core/component/interface/component/types.ts
@@ -6,6 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+import type { State } from 'core/component/state';
+import type { CreateAppFunction } from 'core/component/engines';
 import type { ComponentInterface } from 'core/component/interface/component';
 
 /**
@@ -21,6 +23,14 @@ export interface ComponentConstructor<T = unknown> {
 export type ComponentElement<T = ComponentInterface> = Element & {
 	component?: T;
 };
+
+/**
+ * References to the context of the entire application and the tied state
+ */
+export interface ComponentApp {
+	context: ReturnType<CreateAppFunction>;
+	state: State;
+}
 
 export interface ComponentEmitterOptions {
 	/**