+++ CHANGELOG.md
@@ -122,6 +122,13 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## 4.0.0-beta.71 (2024-03-12)
 
+#### :boom: Breaking Change
+
+* The modules `core/init` and `core/component/state` have been completely redesigned for the new API
+* The field `locale` has been removed `iStaticPage`
+
+## v4.0.0-beta.?? (2024-??-??)
+
 #### :rocket: New Feature
 
 * Created a new global component `g-slider` that encapsulates CSS `scroll snap` logic `gSlider`